pyproject.toml
[tool.poetry]
name = "rasi-chatbot-llm"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
uvicorn = "^0.23.2"
pydantic = ">=2.1.1,<3.0.0"
langchain-openai = "^0.1.8"
langchain-community = "^0.2.4"
langchainhub = "^0.1.18"
python-dotenv = "^1.0.1"
langgraph = "^0.0.66"
tiktoken = "^0.7.0"
unstructured = "^0.14.5"
langchain-text-splitters = "^0.2.1"
weaviate-client = "^4.6.4"
psycopg2 = "^2.9.9"
psycopg2-binary = "^2.9.9"
python-multipart = "^0.0.9"
jq = "^1.7.0"
networkx = "^3.3"
pandas = "^2.2.2"
xlrd = "^2.0.1"
python-pptx = "^0.6.23"
python-docx = "^1.1.2"
docx2txt = "^0.8"
openpyxl = "^3.1.4"
aiosqlite = "^0.20.0"
supabase = "^2.5.1"
fastapi = "^0.111.0"
langchain-qdrant = "^0.1.1"
fastembed = "^0.3.1"
langfuse = "^2.36.2"
psycopg = "^3.2.1"
psycopg-pool = "^3.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
